Move types/versions up a directory in data-types for less nesting
