+++ packages/data-types/src/types/v1/boundary.ts
@@ -1,6 +1,6 @@
 import { FieldType } from 'xlucene-evaluator';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../../interfaces';
+import { ElasticSearchTypes } from '../../interfaces';
 
 export default class Boundary extends BaseType {
     toESMapping(_version?: number) {
