+++ packages/data-types/src/types/v1/integer.ts
@@ -1,7 +1,6 @@
 import { FieldType } from 'xlucene-evaluator';
-
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../../interfaces';
+import { ElasticSearchTypes } from '../../interfaces';
 
 export default class Integer extends BaseType {
     toESMapping(_version?: number) {
