+++ packages/data-types/src/types/v1/keyword-path-analyzer.ts
@@ -1,0 +1,41 @@
+import { FieldType } from 'xlucene-evaluator';
+import BaseType from '../base-type';
+import { ElasticSearchTypes } from '../../interfaces';
+
+export default class KeywordPathAnalyzer extends BaseType {
+    toESMapping(_version?: number) {
+        return {
+            mapping: {
+                [this.field]: {
+                    type: 'keyword' as ElasticSearchTypes,
+                    fields: {
+                        tokens: {
+                            type: 'text',
+                            analyzer: 'path_analyzer',
+                        },
+                    },
+                },
+            },
+            analyzer: {
+                path_analyzer: {
+                    type: 'custom',
+                    tokenizer: 'path_tokenizer'
+                }
+            },
+            tokenizer: {
+                path_tokenizer: {
+                    type: 'pattern',
+                    pattern: '/'
+                }
+            },
+        };
+    }
+
+    toGraphQL() {
+        return this._formatGql('String');
+    }
+
+    toXlucene() {
+        return { [this.field]: FieldType.String };
+    }
+}
