+++ packages/data-types/src/types/v1/keyword-tokens-case-insensitive.ts
@@ -1,8 +1,8 @@
 import { FieldType } from 'xlucene-evaluator';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../../interfaces';
+import { ElasticSearchTypes } from '../../interfaces';
 
-export default class Hostname extends BaseType {
+export default class KeywordTokensCaseInsensitive extends BaseType {
     toESMapping(_version?: number) {
         return {
             mapping: {
@@ -12,26 +12,16 @@ export default class Hostname extends BaseType {
                     fields: {
                         tokens: {
                             type: 'text' as ElasticSearchTypes,
-                            analyzer: 'hostname_analyzer',
+                            analyzer: 'standard',
                         },
                     },
                 },
             },
             analyzer: {
-                hostname_analyzer: {
-                    type: 'custom',
-                    tokenizer: 'hostname_tokenizer'
-                },
                 lowercase_keyword_analyzer: {
                     tokenizer: 'keyword',
                     filter: 'lowercase',
-                }
-            },
-            tokenizer: {
-                hostname_tokenizer: {
-                    type: 'pattern',
-                    pattern: '\\.'
-                }
+                },
             },
         };
     }
