+++ packages/data-types/src/types/v1/long.ts
@@ -1,10 +1,10 @@
 import { FieldType } from 'xlucene-evaluator';
 import BaseType from '../base-type';
-import { ElasticSearchTypes } from '../../../interfaces';
+import { ElasticSearchTypes } from '../../interfaces';
 
-export default class Byte extends BaseType {
+export default class Long extends BaseType {
     toESMapping(_version?: number) {
-        return { mapping: { [this.field]: { type: 'byte' as ElasticSearchTypes } } };
+        return { mapping: { [this.field]: { type: 'long' as ElasticSearchTypes } } };
     }
 
     toGraphQL() {
