+++ packages/data-types/test/types/v1/date-spec.ts
@@ -1,5 +1,5 @@
 import { TSError } from '@terascope/utils';
-import DateType from '../../../src/types/versions/v1/date';
+import DateType from '../../../src/types/v1/date';
 import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('Date V1', () => {
