+++ packages/data-types/test/types/v1/domain-spec.ts
@@ -1,5 +1,5 @@
 import { TSError } from '@terascope/utils';
-import Domain from '../../../src/types/versions/v1/domain';
+import Domain from '../../../src/types/v1/domain';
 import { FieldTypeConfig, ElasticSearchTypes } from '../../../src/interfaces';
 
 describe('Domain V1', () => {
