+++ packages/data-types/test/types/v1/float-spec.ts
@@ -1,5 +1,5 @@
 import { TSError } from '@terascope/utils';
-import FloatType from '../../../src/types/versions/v1/float';
+import FloatType from '../../../src/types/v1/float';
 import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('Float V1', () => {
