+++ packages/data-types/test/types/v1/geo-json-spec.ts
@@ -1,5 +1,5 @@
 import { TSError } from '@terascope/utils';
-import GeoJSONType from '../../../src/types/versions/v1/geo-json';
+import GeoJSONType from '../../../src/types/v1/geo-json';
 import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('GeoPoint V1', () => {
