+++ packages/data-types/test/types/v1/hostname-spec.ts
@@ -1,5 +1,5 @@
 import { TSError } from '@terascope/utils';
-import Hostname from '../../../src/types/versions/v1/hostname';
+import Hostname from '../../../src/types/v1/hostname';
 import { FieldTypeConfig, ElasticSearchTypes } from '../../../src/interfaces';
 
 describe('Hostname V1', () => {
