+++ packages/data-types/test/types/v1/integer-spec.ts
@@ -1,6 +1,6 @@
 import { TSError } from '@terascope/utils';
 import { FieldType } from 'xlucene-evaluator';
-import Interger from '../../../src/types/versions/v1/integer';
+import Interger from '../../../src/types/v1/integer';
 import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('Integer V1', () => {
