+++ packages/data-types/test/types/v1/long-spec.ts
@@ -1,6 +1,6 @@
 import { TSError } from '@terascope/utils';
 import { FieldType } from 'xlucene-evaluator';
-import LongType from '../../../src/types/versions/v1/long';
+import LongType from '../../../src/types/v1/long';
 import { FieldTypeConfig } from '../../../src/interfaces';
 
 describe('Long V1', () => {
